🐛 fix: validating whether the company exists before deleting

+++ packages/core/src/use-cases/companies/delete-company-use-case.ts
@@ -12,6 +12,9 @@ export class DeleteCompanyUseCase {
     }
 
     async execute({ companyId }: Request) {
-        await this.companiesRepository.delete(companyId)
+        const company = await this.companiesRepository.findById(companyId)
+        if (company) {
+            await this.companiesRepository.delete(companyId)
+        }
     }
 }
